src/PyFace/requirements.txt
opencv-python
dlib
grpcio
protobuf